refactor: unify CLI/Just Bash help + arg parsing in one renderer
Single source of truth for generated-client help text and argument
validation, shared across all host transform modes.

- New cli::help module renders top-level help and rich per-subcommand
  help. Differences between the generated CLI --help, the *_help tool
  description, and Just Bash are expressed only as HelpFraming
  (prefix/footer); the body is now byte-identical.
- cli.rs and ffi/host_transform.rs delegate to cli::help (removing the
  duplicated short_tool_description/first_sentence/truncate_clean and
  divergent top-level renderers).
- Native parser (cli/parser.rs) gains JSON Schema enum validation so the
  Just Bash path validates enums like the generated CLI script does.
- Expose render_cli_top_level_help / render_cli_subcommand_help over FFI
  and the napi addon; just_bash_commands.ts now uses them and intercepts
  per-subcommand --help to emit the same rich help as CLI mode.
- Parity tests: CLI body == *_help body (modulo framing); Just Bash
  top-level + subcommand help == generated CLI; Just Bash enum
  validation. Goldens reflect dynamic-width subcommand columns.

CLI-mode behavior is preserved; Just Bash catches up to full parity.
