+++ crates/mcp-compressor-core/src/cli/mod.rs
@@ -1,2 +1,3 @@
+pub mod help;
 pub mod mapping;
 pub mod parser;
