+++ crates/mcp-compressor-core/src/client_gen/cli.rs
@@ -8,10 +8,9 @@
 //! - Passes `Authorization: Bearer <session token>` on every `POST /exec` request.
 //! - Is marked executable (Unix `chmod 755`).
 
-use std::collections::HashSet;
-
 use std::net::ToSocketAddrs;
 
+use crate::cli::help::{self, HelpFraming};
 use crate::cli::mapping::tool_name_to_subcommand;
 use crate::client_gen::generator::{
     CliBridgeEntry, ClientGenerator, GeneratedArtifact, GeneratorConfig,
@@ -43,7 +42,12 @@ fn render_cli_artifacts(config: &GeneratorConfig) -> Vec<GeneratedArtifact> {
 }
 
 fn render_unix_script(config: &GeneratorConfig) -> String {
-    let help = render_top_level_help(config);
+    let help = help::render_top_level_help(
+        &config.cli_name,
+        &config.cli_name,
+        &config.tools,
+        &HelpFraming::shell(&config.cli_name),
+    );
     let bridges = bridge_map_literal(config);
     let mut script = format!(
         r#"#!/usr/bin/env sh
@@ -111,7 +115,7 @@ case "$1" in
 
     for tool in &config.tools {
         let subcommand = tool_name_to_subcommand(&tool.name);
-        let tool_help = render_tool_help(&config.cli_name, tool);
+        let tool_help = help::render_subcommand_help(&config.cli_name, tool);
         script.push_str(&format!(
             r#"  {subcommand})
     shift
@@ -404,388 +408,6 @@ fn bridge_is_live(bridge_url: &str) -> bool {
         })
 }
 
-fn render_top_level_help(config: &GeneratorConfig) -> String {
-    let mut help = format!(
-        "{name} - the {name} toolset\n\nWhen relevant, outputs from this CLI will prefer using the TOON format for more efficient representation of data.\n\nUSAGE:\n  {name} <subcommand> [options]\n\nSUBCOMMANDS:\n",
-        name = config.cli_name,
-    );
-    for tool in &config.tools {
-        help.push_str(&format!(
-            "  {:<35} {}\n",
-            tool_name_to_subcommand(&tool.name),
-            short_tool_description(tool.description.as_deref()),
-        ));
-    }
-    help.push_str(&format!(
-        "\nRun '{} <subcommand> --help' for subcommand usage.",
-        config.cli_name
-    ));
-    help
-}
-
-fn render_tool_help(cli_name: &str, tool: &crate::compression::engine::Tool) -> String {
-    let subcommand = tool_name_to_subcommand(&tool.name);
-    let mut help = format!(
-        "{cli_name} {subcommand}\n\n{}\n\nUSAGE:\n  {cli_name} {subcommand} [options]\n",
-        tool_description_block(tool.description.as_deref().unwrap_or("Invoke this tool.")),
-    );
-
-    let options = tool_options(tool);
-    let required = options
-        .iter()
-        .filter(|option| option.required && !option.is_complex)
-        .collect::<Vec<_>>();
-    let optional = options
-        .iter()
-        .filter(|option| !option.required && !option.is_complex)
-        .collect::<Vec<_>>();
-    let complex = options
-        .iter()
-        .filter(|option| option.is_complex)
-        .collect::<Vec<_>>();
-
-    append_option_section(&mut help, "REQUIRED", &required);
-    append_option_section(&mut help, "OPTIONAL", &optional);
-    append_option_section(&mut help, "COMPLEX / JSON OPTIONS", &complex);
-
-    help.push_str("\nGLOBAL OPTIONS:\n");
-    help.push_str("  --json <json object>\n");
-    help.push_str("      Provide the entire tool input as JSON. This bypasses flag parsing.\n");
-    help.push_str("\n  --help\n");
-    help.push_str("      Show this help message.\n");
-
-    help
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ToolOption {
-    name: String,
-    ty: String,
-    required: bool,
-    description: Option<String>,
-    default: Option<String>,
-    enum_values: Vec<String>,
-    minimum: Option<String>,
-    maximum: Option<String>,
-    min_length: Option<String>,
-    max_length: Option<String>,
-    min_items: Option<String>,
-    max_items: Option<String>,
-    is_complex: bool,
-    complex_hint: Option<String>,
-}
-
-fn tool_options(tool: &crate::compression::engine::Tool) -> Vec<ToolOption> {
-    let required = tool
-        .input_schema
-        .get("required")
-        .and_then(|value| value.as_array())
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(|value| value.as_str().map(str::to_string))
-                .collect::<HashSet<_>>()
-        })
-        .unwrap_or_default();
-
-    tool.input_schema
-        .get("properties")
-        .and_then(|value| value.as_object())
-        .map(|properties| {
-            properties
-                .iter()
-                .map(|(name, schema)| {
-                    let (ty, is_complex, complex_hint) = schema_type_details(schema);
-                    ToolOption {
-                        name: name.clone(),
-                        ty,
-                        required: required.contains(name),
-                        description: schema
-                            .get("description")
-                            .and_then(|value| value.as_str())
-                            .map(full_description),
-                        default: schema.get("default").map(default_value_label),
-                        enum_values: schema_enum_values(schema),
-                        minimum: schema_number_constraint(schema, "minimum"),
-                        maximum: schema_number_constraint(schema, "maximum"),
-                        min_length: schema_number_constraint(schema, "minLength"),
-                        max_length: schema_number_constraint(schema, "maxLength"),
-                        min_items: schema_number_constraint(schema, "minItems"),
-                        max_items: schema_number_constraint(schema, "maxItems"),
-                        is_complex,
-                        complex_hint,
-                    }
-                })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn append_option_section(help: &mut String, title: &str, options: &[&ToolOption]) {
-    if options.is_empty() {
-        return;
-    }
-    help.push('\n');
-    help.push_str(title);
-    help.push_str(":\n");
-    for option in options {
-        help.push_str(&format_tool_option_help(option));
-    }
-}
-
-fn format_tool_option_help(option: &ToolOption) -> String {
-    let flag = format!("--{}", tool_name_to_subcommand(&option.name));
-    let mut output = format!("  {flag} <{}>\n", option.ty);
-    let mut details = Vec::new();
-    if option.required {
-        details.push("Required.".to_string());
-    }
-    if let Some(description) = &option.description {
-        details.push(description.clone());
-    }
-    if !option.enum_values.is_empty() {
-        details.push(format!(
-            "Allowed values: {}.",
-            option.enum_values.join(", ")
-        ));
-    }
-    if let Some(default) = &option.default {
-        details.push(format!("Default: {default}."));
-    }
-    if let Some(minimum) = &option.minimum {
-        details.push(format!("Minimum: {minimum}."));
-    }
-    if let Some(maximum) = &option.maximum {
-        details.push(format!("Maximum: {maximum}."));
-    }
-    if let Some(min_length) = &option.min_length {
-        details.push(format!("Minimum length: {min_length}."));
-    }
-    if let Some(max_length) = &option.max_length {
-        details.push(format!("Maximum length: {max_length}."));
-    }
-    if let Some(min_items) = &option.min_items {
-        details.push(format!("Minimum items: {min_items}."));
-    }
-    if let Some(max_items) = &option.max_items {
-        details.push(format!("Maximum items: {max_items}."));
-    }
-    if option.ty == "boolean" {
-        details.push("Accepted values: true, false.".to_string());
-        details.push(format!(
-            "Also supports: {flag} and --no-{}.",
-            tool_name_to_subcommand(&option.name)
-        ));
-    }
-    if let Some(hint) = &option.complex_hint {
-        details.push(hint.clone());
-    }
-    for detail in details {
-        output.push_str(&wrap_indented(&detail, 6, 100));
-    }
-    output
-}
-
-fn schema_number_constraint(schema: &serde_json::Value, key: &str) -> Option<String> {
-    schema.get(key).map(default_value_label)
-}
-
-fn schema_type_details(schema: &serde_json::Value) -> (String, bool, Option<String>) {
-    let labels = schema_enum_values(schema);
-    if !labels.is_empty() {
-        return (labels.join("|"), false, None);
-    }
-    if schema.get("oneOf").is_some()
-        || schema.get("anyOf").is_some()
-        || schema.get("allOf").is_some()
-    {
-        return (
-            "json".to_string(),
-            true,
-            Some("Schema contains oneOf/anyOf/allOf; use --json for complex input.".to_string()),
-        );
-    }
-    match schema.get("type").and_then(|value| value.as_str()) {
-        Some("integer") => ("integer".to_string(), false, None),
-        Some("number") => ("number".to_string(), false, None),
-        Some("boolean") => ("boolean".to_string(), false, None),
-        Some("array") => match schema.get("items") {
-            Some(items) => {
-                let (item_ty, item_complex, _) = schema_type_details(items);
-                if item_complex
-                    || matches!(
-                        items.get("type").and_then(|value| value.as_str()),
-                        Some("object" | "array")
-                    )
-                {
-                    (
-                        "json array".to_string(),
-                        true,
-                        Some("Pass as a JSON array string.".to_string()),
-                    )
-                } else {
-                    (
-                        format!("{item_ty}[]"),
-                        false,
-                        Some("Repeat this flag or pass a JSON array.".to_string()),
-                    )
-                }
-            }
-            None => (
-                "json array".to_string(),
-                true,
-                Some("Pass as a JSON array string.".to_string()),
-            ),
-        },
-        Some("object") => (
-            "json object".to_string(),
-            true,
-            Some("Pass as a JSON object string.".to_string()),
-        ),
-        Some("string") | None => ("string".to_string(), false, None),
-        Some(other) => (other.to_string(), false, None),
-    }
-}
-
-fn full_description(description: &str) -> String {
-    description.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-/// Render a tool's top-level description while preserving its original line
-/// structure. Each line is trimmed and has internal whitespace runs collapsed,
-/// but newlines (including blank lines between paragraphs) are kept so the
-/// description does not get flattened into a single massive line. Leading and
-/// trailing blank lines are removed, and runs of 3+ blank lines are collapsed
-/// to a single blank line.
-fn tool_description_block(description: &str) -> String {
-    let mut lines: Vec<String> = description
-        .lines()
-        .map(|line| line.split_whitespace().collect::<Vec<_>>().join(" "))
-        .collect();
-
-    // Collapse multiple consecutive blank lines into a single blank line.
-    let mut collapsed: Vec<String> = Vec::with_capacity(lines.len());
-    let mut previous_blank = false;
-    for line in lines.drain(..) {
-        let is_blank = line.is_empty();
-        if is_blank && previous_blank {
-            continue;
-        }
-        previous_blank = is_blank;
-        collapsed.push(line);
-    }
-
-    // Trim leading/trailing blank lines.
-    while collapsed.first().is_some_and(|line| line.is_empty()) {
-        collapsed.remove(0);
-    }
-    while collapsed.last().is_some_and(|line| line.is_empty()) {
-        collapsed.pop();
-    }
-
-    if collapsed.is_empty() {
-        return "Invoke this tool.".to_string();
-    }
-
-    collapsed.join("\n")
-}
-
-fn wrap_indented(value: &str, indent: usize, width: usize) -> String {
-    let prefix = " ".repeat(indent);
-    let mut output = String::new();
-    let mut line = String::new();
-    for word in value.split_whitespace() {
-        if !line.is_empty() && line.len() + 1 + word.len() > width.saturating_sub(indent) {
-            output.push_str(&prefix);
-            output.push_str(line.trim_end());
-            output.push('\n');
-            line.clear();
-        }
-        if !line.is_empty() {
-            line.push(' ');
-        }
-        line.push_str(word);
-    }
-    if !line.is_empty() {
-        output.push_str(&prefix);
-        output.push_str(line.trim_end());
-        output.push('\n');
-    }
-    output
-}
-
-fn schema_enum_values(schema: &serde_json::Value) -> Vec<String> {
-    schema
-        .get("enum")
-        .and_then(|value| value.as_array())
-        .map(|values| values.iter().map(default_value_label).collect())
-        .unwrap_or_default()
-}
-
-fn default_value_label(value: &serde_json::Value) -> String {
-    match value {
-        serde_json::Value::String(value) => value.clone(),
-        other => other.to_string(),
-    }
-}
-
-fn short_tool_description(description: Option<&str>) -> String {
-    let trimmed = description.unwrap_or_default().trim();
-    if trimmed.is_empty() {
-        return String::new();
-    }
-
-    let first_sentence = first_sentence(trimmed);
-    let candidate = if first_sentence.chars().count() >= 10 {
-        first_sentence
-    } else {
-        first_non_empty_line(trimmed)
-    };
-    truncate_clean(candidate, 200)
-}
-
-fn first_sentence(value: &str) -> &str {
-    for (index, ch) in value.char_indices() {
-        if matches!(ch, '.' | '!' | '?') {
-            return value[..=index].trim();
-        }
-    }
-    first_non_empty_line(value)
-}
-
-fn first_non_empty_line(value: &str) -> &str {
-    value
-        .lines()
-        .map(str::trim)
-        .find(|line| !line.is_empty())
-        .unwrap_or_default()
-}
-
-fn truncate_clean(value: &str, max_chars: usize) -> String {
-    let compact = value.split_whitespace().collect::<Vec<_>>().join(" ");
-    if compact.chars().count() <= max_chars {
-        return compact;
-    }
-    let limit = max_chars.saturating_sub(3);
-    let mut end = 0;
-    for (count, (index, ch)) in compact.char_indices().enumerate() {
-        if count >= limit {
-            break;
-        }
-        end = index + ch.len_utf8();
-    }
-    let mut prefix = compact[..end]
-        .trim_end_matches(|ch: char| ch.is_whitespace() || ch == ',' || ch == ';' || ch == ':')
-        .to_string();
-    if let Some(space) = prefix.rfind(' ') {
-        if space >= max_chars / 2 {
-            prefix.truncate(space);
-        }
-    }
-    prefix.push_str("...");
-    prefix
-}
-
 fn shell_quote(value: &str) -> String {
     value.replace('\'', "'\\''")
 }
@@ -989,21 +611,6 @@ mod tests {
         assert!(content.contains("Default: 30."));
         assert!(content.contains("--method"));
         assert!(content.contains("Allowed values: GET, POST."));
-    }
-
-    /// The top-level tool description preserves its multi-line structure rather
-    /// than being flattened into one long line, and drops the noisy OUTPUT
-    /// section.
-    #[test]
-    fn tool_description_block_preserves_line_structure() {
-        let input = "  First line.  \n\nSecond  paragraph    with   spaces.\nThird line.\n\n\n";
-        let rendered = tool_description_block(input);
-        assert_eq!(
-            rendered,
-            "First line.\n\nSecond paragraph with spaces.\nThird line."
-        );
-        // Multi-line descriptions must keep newlines (not collapse to one line).
-        assert!(rendered.contains('\n'));
     }
 
     /// Subcommand help no longer emits the redundant OUTPUT section.
