+++ crates/mcp-compressor-core/src/ffi/mod.rs
@@ -17,14 +17,14 @@ pub use client_gen::{
     generate_client_artifact_files, generate_client_artifacts, maybe_toonify_output,
     render_client_artifacts, FfiClientArtifactKind,
 };
-pub use host_transform::{
-    build_host_transform_plan, normalize_host_tool_result, FfiHostJustBashCommandPlan,
-    FfiHostJustBashPlan, FfiHostTransformConfig, FfiHostTransformKind, FfiHostTransformPlan,
-};
 pub use dto::{
     FfiBackendConfig, FfiCompressedSessionConfig, FfiCompressedSessionInfo, FfiGeneratorConfig,
     FfiJustBashCommandSpec, FfiJustBashProviderSpec, FfiMcpServer, FfiSdkServerConfig,
     FfiSdkServersConfig, FfiTool,
+};
+pub use host_transform::{
+    build_host_transform_plan, normalize_host_tool_result, FfiHostJustBashCommandPlan,
+    FfiHostJustBashPlan, FfiHostTransformConfig, FfiHostTransformKind, FfiHostTransformPlan,
 };
 pub use oauth::{
     clear_oauth_credentials, list_oauth_credentials, oauth_store_path, remember_oauth_backend,
@@ -32,6 +32,7 @@ pub use oauth::{
 };
 pub use pure::{
     compress_tool_listing, format_tool_schema_response, parse_mcp_config, parse_tool_argv,
+    render_cli_subcommand_help, render_cli_top_level_help,
 };
 pub use session::{
     normalize_sdk_servers, start_compressed_session, start_compressed_session_from_mcp_config,
