+++ crates/mcp-compressor-core/tests/generated_clients.rs
@@ -104,7 +104,10 @@ fn generated_cli_help_renders_camel_case_properties_as_kebab_case_flags() {
     let stdout = String::from_utf8_lossy(&output.stdout);
     assert!(stdout.contains("--cloud-id <string>"), "stdout: {stdout}");
     assert!(stdout.contains("--jql <string>"), "stdout: {stdout}");
-    assert!(stdout.contains("--max-results <number>"), "stdout: {stdout}");
+    assert!(
+        stdout.contains("--max-results <number>"),
+        "stdout: {stdout}"
+    );
     assert!(
         stdout.contains("--next-page-token <string>"),
         "stdout: {stdout}"
